refactor(utils): rename `util` submodule files
