+++ src/declusor/interface/__init__.py
@@ -1,4 +1,4 @@
-from .parser import IParser, DeclusorArguments
+from .parser import DeclusorArguments, IParser
 from .prompt import IPrompt
 from .router import Controller, IRouter
 from .session import ISession
