+++ src/declusor/util/__init__.py
@@ -1,10 +1,10 @@
 from .client import format_client_script, format_function_call
 from .encoding import convert_bytes_to_hex
-from .filesystem import load_file, load_library, load_payload
 from .message import read_stripped_message, write_binary_data, write_error_message, write_string_message, write_warning_message
+from .network import await_connection
 from .parsing import parse_command_arguments
-from .sanitize import validate_file_extension, validate_file_relative
-from .socket import await_connection
+from .security import validate_file_extension, validate_file_relative
+from .storage import load_file, load_library, load_payload
 
 __all__ = [
     "await_connection",
