+++ tests/test_util_parsing.py
@@ -1,7 +1,7 @@
 import unittest
 
 from declusor.config import ArgumentParsingError
-from declusor.core import parse_command_arguments
+from declusor.util import parsing
 
 
 class TestParseUtil(unittest.TestCase):
@@ -11,7 +11,7 @@ class TestParseUtil(unittest.TestCase):
         line = "arg1"
         definitions = {"arg": str}
 
-        parsed, unknown = parse_command_arguments(line, definitions)
+        parsed, unknown = parsing.parse_command_arguments(line, definitions)
 
         self.assertEqual(parsed, {"arg": "arg1"})
         self.assertEqual(unknown, [])
@@ -22,7 +22,7 @@ class TestParseUtil(unittest.TestCase):
         line = "val1 123"
         definitions: dict[str, type] = {"str_arg": str, "int_arg": int}
 
-        parsed, unknown = parse_command_arguments(line, definitions)
+        parsed, unknown = parsing.parse_command_arguments(line, definitions)
 
         self.assertEqual(parsed, {"str_arg": "val1", "int_arg": 123})
 
@@ -32,7 +32,7 @@ class TestParseUtil(unittest.TestCase):
         line = "val1 extra1 extra2"
         definitions = {"arg": str}
 
-        parsed, unknown = parse_command_arguments(line, definitions, allow_unknown=True)
+        parsed, unknown = parsing.parse_command_arguments(line, definitions, allow_unknown=True)
 
         self.assertEqual(parsed, {"arg": "val1"})
         self.assertEqual(unknown, ["extra1", "extra2"])
@@ -44,7 +44,7 @@ class TestParseUtil(unittest.TestCase):
         definitions = {"arg": str}
 
         with self.assertRaises(ArgumentParsingError):
-            parse_command_arguments(line, definitions)
+            parsing.parse_command_arguments(line, definitions)
 
     def test_parse_error_unknown_arg_not_allowed(self) -> None:
         """Test that unknown arguments raise an error when not allowed."""
@@ -53,12 +53,12 @@ class TestParseUtil(unittest.TestCase):
         definitions = {"arg": str}
 
         with self.assertRaises(ArgumentParsingError):
-            parse_command_arguments(line, definitions, allow_unknown=False)
+            parsing.parse_command_arguments(line, definitions, allow_unknown=False)
 
     def test_parse_empty_no_defs(self) -> None:
         """Test parsing empty line with no definitions."""
 
-        parsed, unknown = parse_command_arguments("", {})
+        parsed, unknown = parsing.parse_command_arguments("", {})
 
         self.assertEqual(parsed, {})
         self.assertEqual(unknown, [])
